backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
